chore: relocate annotations to core package
