+++ fullstack-junit-support/src/main/java/module-info.java
@@ -1,6 +1,6 @@
 module com.hedera.fullstack.junit.support {
     exports com.hedera.fullstack.junit.support;
-    exports com.hedera.fullstack.junit.support.annotation;
+    exports com.hedera.fullstack.junit.support.annotation.core;
     exports com.hedera.fullstack.junit.support.annotation.node;
     exports com.hedera.fullstack.junit.support.annotation.services;
     exports com.hedera.fullstack.junit.support.annotation.validation;
